@@ -29,8 +29,8 @@ project.ext.limits = [
         'class'      : 90
 ]
 
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
+sourceCompatibility = JavaVersion.VERSION_11
+targetCompatibility = JavaVersion.VERSION_11
 compileJava.options.encoding = 'UTF-8'
 compileTestJava.options.encoding = 'UTF-8'
 
@@ -39,19 +39,19 @@ repositories {
 }
 
 dependencies {
-    api 'com.epam.reportportal:client-java:5.3.17'
+    api 'com.epam.reportportal:client-java:5.4.3'
     api 'com.nordstrom.tools:junit-foundation:17.2.4'
 
     implementation 'org.slf4j:slf4j-api:2.0.7'
 
-    testImplementation 'com.epam.reportportal:agent-java-test-utils:0.0.13'
+    testImplementation 'com.epam.reportportal:agent-java-test-utils:0.1.0'
 
     testImplementation 'org.aspectj:aspectjweaver:1.9.19'
     testImplementation 'org.hamcrest:hamcrest-core:2.2'
     testImplementation "org.mockito:mockito-core:${project.mockito_version}"
     testImplementation "org.mockito:mockito-junit-jupiter:${project.mockito_version}"
-    testImplementation 'ch.qos.logback:logback-classic:1.3.15'
-    testImplementation 'com.epam.reportportal:logger-java-logback:5.2.3'
+    testImplementation 'ch.qos.logback:logback-classic:1.5.18'
+    testImplementation 'com.epam.reportportal:logger-java-logback:5.4.0'
     testImplementation ("org.junit.platform:junit-platform-runner:${project.junit5_runner_version}") {
         exclude module: 'junit'
     }
