@@ -21,14 +21,16 @@ import com.epam.reportportal.junit.utils.TestUtils;
 import com.epam.reportportal.service.ReportPortal;
 import com.epam.reportportal.service.ReportPortalClient;
 import com.epam.reportportal.util.test.CommonUtils;
+import com.epam.ta.reportportal.ws.model.ApiInfo;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
+import io.reactivex.Maybe;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatchers;
 
-import java.util.Date;
+import java.time.Instant;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -62,6 +64,11 @@ public class ItemTimeOrderTest {
 		TestUtils.mockLaunch(client, null, firstSuiteId, secondSuiteId, classIds);
 		TestUtils.mockNestedSteps(client, tests);
 		TestUtils.mockBatchLogging(client);
+		ApiInfo info = new ApiInfo();
+		ApiInfo.Build build = new ApiInfo.Build();
+		info.setBuild(build);
+		build.setVersion("5.13.2");
+		when(client.getApiInfo()).thenReturn(Maybe.just(info));
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters(), TestUtils.testExecutor()));
 	}
 
@@ -82,17 +89,20 @@ public class ItemTimeOrderTest {
 		verify(client, timeout(PROCESSING_TIMEOUT)).startTestItem(same(classIds.get(0)), stepCaptor.capture());
 		verify(client, timeout(PROCESSING_TIMEOUT)).startTestItem(same(classIds.get(1)), stepCaptor.capture());
 
-		Date previousDate = null;
+		Instant previousDate = null;
 		List<StartTestItemRQ> testItems = testCaptor.getAllValues();
 		List<StartTestItemRQ> suites = Stream.concat(suiteCaptor.getAllValues().stream(), testItems.stream()).collect(Collectors.toList());
 		for (StartTestItemRQ item : suites) {
-			Date parentDate = ofNullable(previousDate).orElseGet(() -> suites.get(0).getStartTime());
-			Date itemDate = item.getStartTime();
-			assertThat(item.getStartTime(), greaterThanOrEqualTo(parentDate));
+			Instant parentDate = ofNullable(previousDate).orElseGet(() -> (Instant) suites.get(0).getStartTime());
+			Instant itemDate = (Instant) item.getStartTime();
+			assertThat((Instant) item.getStartTime(), greaterThanOrEqualTo(parentDate));
 			previousDate = itemDate;
 		}
 		List<StartTestItemRQ> steps = stepCaptor.getAllValues();
 		IntStream.range(0, TEST_NUMBER)
-				.forEach(i -> assertThat(steps.get(i).getStartTime(), greaterThanOrEqualTo(testItems.get(i).getStartTime())));
+				.forEach(i -> assertThat(
+						(Instant) steps.get(i).getStartTime(),
+						greaterThanOrEqualTo((Instant) testItems.get(i).getStartTime())
+				));
 	}
 }
