@@ -42,7 +42,8 @@ public class SystemAttributesFetcher {
 	}
 
 	public static Set<ItemAttributesRQ> collectSystemAttributes(Boolean skippedAnIssue) {
-		Set<ItemAttributesRQ> systemAttributes = SystemAttributesExtractor.extract(PROPS_FILE,
+		Set<ItemAttributesRQ> systemAttributes = SystemAttributesExtractor.extract(
+				PROPS_FILE,
 				SystemAttributesFetcher.class.getClassLoader()
 		);
 		systemAttributes.add(skippedIssue(skippedAnIssue));
